% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{compare_boxplots}
\alias{compare_boxplots}
\title{compares the prs for the cases and controls}
\usage{
compare_boxplots(full_table, phenotype, score)
}
\arguments{
\item{full_table}{data.frame containing phenotype and score information}

\item{phenotype}{character vector for column name for the phenotype}

\item{score}{character vector for column name for score}
}
\description{
compares the prs for the cases and controls
}
