% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{join_scores}
\alias{join_scores}
\title{join phenotype file with score file}
\usage{
join_scores(pheno_table, score_table)
}
\arguments{
\item{pheno_table}{data.frame containing phenotypic information}

\item{score_table}{data.frame containing score information}
}
\description{
join phenotype file with score file
}
