% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{k_fold}
\alias{k_fold}
\title{performs 5-fold cross validation}
\usage{
k_fold(phenotype, score_col, full_table)
}
\arguments{
\item{phenotype}{character vector for phenotype}

\item{score_col}{character vector for score}

\item{full_table}{data.frame containing phenotype and score information}
}
\description{
performs 5-fold cross validation
}
