% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno_data}
\alias{pheno_data}
\title{phenotype table
a mock dataset containing phenotype information}
\format{data.frame with 3300 rows and 6 columns
\describe{
  \item{hypertension}{phenotype status, 1 or 0}
  \item{age}{individual's age group, by decade}
  \item{sex}{individual's sex, 1 denotes male, 0 denotes female}
  \item{IID}{individual identifier for each subject}
  \item{diabetes}{phenotype status, 1 or 0}
  \item{ms}{phenotype status, 1 or 0}
}}
\usage{
pheno_data
}
\description{
phenotype table
a mock dataset containing phenotype information
}
\keyword{datasets}
