% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{pheno_select}
\alias{pheno_select}
\title{Selects column from phenotypic dataset to be used in analyses}
\usage{
pheno_select(pheno_table, pheno_name)
}
\arguments{
\item{pheno_table}{data.frame containing phenotypic information}

\item{pheno_name}{character vector column name of the phenotype of interest}
}
\description{
Selects column from phenotypic dataset to be used in analyses
}
