% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{positive_subjects}
\alias{positive_subjects}
\title{outputs the number of subjects who are positive for the phenotype}
\usage{
positive_subjects(full_table, phenotype)
}
\arguments{
\item{full_table}{data.frame containing phenotype and score information}

\item{phenotype}{column name for phenotype}
}
\description{
outputs the number of subjects who are positive for the phenotype
}
