% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{prev_plot}
\alias{prev_plot}
\title{creates a plot that depicts the relationship between PRS decile and phenotypic prevalence in the cohort}
\usage{
prev_plot(full_table, prs_col, pheno_col, prs_num, pheno_num, qtile)
}
\arguments{
\item{full_table}{data.fram containing phenotype and score information}

\item{pheno_col}{character vector corresponding to column name for phenotype}

\item{prs_num}{column number corresponding to score column}

\item{pheno_num}{column number corresponding to phenotype column}

\item{qtile}{number of quantile desired}

\item{PRS_col}{character vector corresponding to column name for scores}
}
\description{
creates a plot that depicts the relationship between PRS decile and phenotypic prevalence in the cohort
}
