% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{proc_analysis}
\alias{proc_analysis}
\title{runs ROC analysis}
\usage{
proc_analysis(full_table, phenotype, my_model, ...)
}
\arguments{
\item{full_table}{data.frame containing phenotype and score information}

\item{phenotype}{character vector for phenotype}

\item{my_model}{glm for Y ~ prs (+ covariates)}

\item{...}{arguments passed}
}
\description{
runs ROC analysis
}
