% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{score_data}
\alias{score_data}
\title{score information
a mock dataset containing score information}
\format{data.frame with 3300 rows and 2 columns
\describe{
  \item{prs}{polygenic risk score}
  \item{IID}{individual identifier for each subject}
}}
\usage{
score_data
}
\description{
score information
a mock dataset containing score information
}
\keyword{datasets}
