% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{score_distribution}
\alias{score_distribution}
\alias{score_distribtuion}
\title{show a histogram of the distribution of scores for the individuals in the cohort}
\usage{
score_distribtuion(full_table, prs)
}
\arguments{
\item{full_table}{data.frame containing phenotype and score information}

\item{prs}{character vector for column name of score column}
}
\description{
show a histogram of the distribution of scores for the individuals in the cohort
}
