% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{simple_logistic_reg}
\alias{simple_logistic_reg}
\alias{complex_logistic_reg}
\title{Runs a simple logistic regression with only prs as a predictor.}
\usage{
simple_logistic_reg(full_table, phenotype, ...)

complex_logistic_reg(full_table, phenotype, ...)
}
\arguments{
\item{full_table}{data.frame containing phenotype and score information}

\item{phenotype}{column name for phenotype}

\item{...}{arguments passed}
}
\description{
Runs a simple logistic regression with only prs as a predictor.

Runs a more complex logistic regression with prs, age and sex as predictors
}
