% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{score and phenotype
a mock dataset containing score and phenotype information}
\format{data.frame with 3300 rows and 4 columns
\describe{
  \item{prs}{polygenic risk score}
  \item{hypertension}{phenotype status, 1 or 0}
  \item{age}{individual's age group, by decade}
  \item{sex}{individual's sex, 1 denotes male, 0 denotes female}
}}
\usage{
test_data
}
\description{
score and phenotype
a mock dataset containing score and phenotype information
}
\keyword{datasets}
