% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{k_fold}
\alias{k_fold}
\title{performs 5-fold cross validation}
\usage{
k_fold(full_table, phenotype, age_col, sex_col, score_col)
}
\arguments{
\item{full_table}{data.frame containing phenotype and score information}

\item{phenotype}{column name for phenotype}

\item{age_col}{age column}

\item{sex_col}{sex column}

\item{score_col}{score column}
}
\description{
performs 5-fold cross validation
}
